+++ src/OnePoleFilter.cpp
@@ -22,6 +22,6 @@ float OnePoleFilter::processSample (float sample)
 
 void OnePoleFilter::setCoefficients (float frequency)
 {
-	m_B1 = exp( -2.0f * M_PI * (frequency / SAMPLE_RATE / 2.0f) );
+	m_B1 = expf( -2.0f * M_PI * (frequency / SAMPLE_RATE / 2.0f) );
 	m_A0 = 1.0f - m_B1;
 }
